Completed tasks for conversion, upgrade and building

+++ Unlimited-NetworkingServer-MiningGame/Headquarters/BuildTask.cs
@@ -7,7 +7,7 @@ namespace Unlimited_NetworkingServer_MiningGame.Headquarters
     /// </summary>
     public class BuildTask : IDarkRiftSerializable
     {
-        public byte Id { get; set; }
+        public ushort Id { get; set; }
         public byte Type { get; set; }
         public byte Element { get; set; }
         public long StartTime { get; set; }
@@ -20,7 +20,7 @@ namespace Unlimited_NetworkingServer_MiningGame.Headquarters
             StartTime = 0;
         }
   
-        public BuildTask(byte id, byte type, byte element, long startTime)
+        public BuildTask(ushort id, byte type, byte element, long startTime)
         {
             Id = id;
             Type = type;
@@ -34,7 +34,7 @@ namespace Unlimited_NetworkingServer_MiningGame.Headquarters
         /// <param name="e">Deserialize event</param>
         public void Deserialize(DeserializeEvent e)
         {
-            Id = e.Reader.ReadByte();
+            Id = e.Reader.ReadUInt16();
             Type = e.Reader.ReadByte();
             Element = e.Reader.ReadByte();
             StartTime = e.Reader.ReadInt64();
